01. Tính năng Phân quyền

+++ controllers/admin/tour.controller.js
@@ -119,6 +119,14 @@ module.exports.create = async (req, res) => {
 };
 
 module.exports.createPost = async (req, res) => {
+  if (!req.permissions.includes("tour-create")) {
+    res.json({
+      code: "error",
+      message: "Không có quyền!",
+    });
+    return;
+  }
+
   if (req.body.position) {
     req.body.position = parseInt(req.body.position);
   } else {
@@ -200,6 +208,14 @@ module.exports.edit = async (req, res) => {
 
 module.exports.editPatch = async (req, res) => {
   try {
+    if (!req.permissions.includes("tour-edit")) {
+      res.json({
+        code: "error",
+        message: "Không có quyền!",
+      });
+      return;
+    }
+
     const id = req.params.id;
     if (req.body.position) {
       req.body.position = parseInt(req.body.position);
@@ -363,6 +379,14 @@ module.exports.trash = async (req, res) => {
 
 module.exports.undoPatch = async (req, res) => {
   try {
+    if (!req.permissions.includes("tour-trash")) {
+      res.json({
+        code: "error",
+        message: "Không có quyền!",
+      });
+      return;
+    }
+
     const id = req.params.id;
 
     await Tour.updateOne(
@@ -388,6 +412,14 @@ module.exports.undoPatch = async (req, res) => {
 
 module.exports.destroyDelete = async (req, res) => {
   try {
+    if (!req.permissions.includes("tour-trash")) {
+      res.json({
+        code: "error",
+        message: "Không có quyền!",
+      });
+      return;
+    }
+
     const id = req.params.id;
 
     await Tour.deleteOne({
@@ -412,6 +444,14 @@ module.exports.changeMultiPatch = async (req, res) => {
     switch (option) {
       case "active":
       case "inactive":
+        if (!req.permissions.includes("tour-edit")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền!",
+          });
+          return;
+        }
+
         await Tour.updateMany(
           {
             _id: { $in: ids },
@@ -426,6 +466,14 @@ module.exports.changeMultiPatch = async (req, res) => {
         });
         break;
       case "undo":
+        if (!req.permissions.includes("tour-trash")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền!",
+          });
+          return;
+        }
+
         await Tour.updateMany(
           {
             _id: { $in: ids },
@@ -440,6 +488,14 @@ module.exports.changeMultiPatch = async (req, res) => {
         });
         break;
       case "delete":
+        if (!req.permissions.includes("tour-trash")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền!",
+          });
+          return;
+        }
+
         await Tour.updateMany(
           {
             _id: { $in: ids },
@@ -456,6 +512,14 @@ module.exports.changeMultiPatch = async (req, res) => {
         });
         break;
       case "destroy":
+        if (!req.permissions.includes("tour-trash")) {
+          res.json({
+            code: "error",
+            message: "Không có quyền!",
+          });
+          return;
+        }
+
         await Tour.deleteMany({
           _id: { $in: ids },
         });
